* Split ZipArchive class to .cpp and .hpp * Add Doxygen documentation to ZipArchive class. * Update MANIFEST file/ * Update CMakeLists.txt file
