+++ src/CMakeLists.txt
@@ -86,6 +86,7 @@ add_library(libslic3r STATIC
     ${LIBDIR}/libslic3r/SurfaceCollection.cpp
     ${LIBDIR}/libslic3r/SVG.cpp
     ${LIBDIR}/libslic3r/TriangleMesh.cpp
+    ${LIBDIR}/libslic3r/Zip/ZipArchive.cpp
 )
 
 add_library(admesh STATIC
